// src/app/modules/logistica/entrada-materiais/status-recebimento/models/statusRecebimento.ts
export interface ILogisticaEntradaMateriaisStatusRecebimento {
    ID_LOGI_ENMA_NOFI_STAT:   number;
    NM_STAT:                  number;
    IN_STAT:                  number; 
    DS_OBSE:                  string;
    NR_MATR:                  number;
    NM_USUA:                  string;
    IN_FHNC:                  string;
    DT_INCL:                  string; 
    DT_ATUA:                  string;
  }
  